Introduce abstract service builder
