+++ server/src/main/java/io/spine/server/QueryService.java
@@ -26,8 +26,6 @@
 package io.spine.server;
 
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Sets;
-import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import io.grpc.stub.StreamObserver;
 import io.spine.client.Query;
 import io.spine.client.QueryResponse;
@@ -38,7 +36,6 @@ import io.spine.server.stand.InvalidRequestException;
 import io.spine.type.TypeUrl;
 
 import java.util.Map;
-import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.flogger.LazyArgs.lazy;
@@ -120,33 +117,11 @@ public final class QueryService
     /**
      * The builder for a {@code QueryService}.
      */
-    public static class Builder {
+    public static class Builder extends AbstractServiceBuilder<QueryService, Builder> {
 
-        private final Set<BoundedContext> contexts = Sets.newHashSet();
-
-        /**
-         * Adds the passed bounded context to be served by the query service.
-         */
-        @CanIgnoreReturnValue
-        public Builder add(BoundedContext context) {
-            contexts.add(context);
+        @Override
+        Builder self() {
             return this;
-        }
-
-        /**
-         * Excludes the passed bounded context from being served by the query service.
-         */
-        @CanIgnoreReturnValue
-        public Builder remove(BoundedContext context) {
-            contexts.remove(context);
-            return this;
-        }
-
-        /**
-         * Tells if the builder already contains the passed bounded context.
-         */
-        public boolean contains(BoundedContext context) {
-            return contexts.contains(context);
         }
 
         /**
@@ -154,8 +129,9 @@ public final class QueryService
          *
          * @throws IllegalStateException if no bounded contexts were added.
          */
+        @Override
         public QueryService build() throws IllegalStateException {
-            if (contexts.isEmpty()) {
+            if (isEmpty()) {
                 var message = "Query service must have at least one `BoundedContext`.";
                 throw new IllegalStateException(message);
             }
@@ -166,7 +142,7 @@ public final class QueryService
 
         private ImmutableMap<TypeUrl, BoundedContext> createMap() {
             ImmutableMap.Builder<TypeUrl, BoundedContext> map = ImmutableMap.builder();
-            for (var context : contexts) {
+            for (var context : contexts()) {
                 putExposedTypes(context, map);
             }
             return map.build();
